Create Day4_PASS.cpp
This helped me achieve my goal of getting a window. This seemingly simple task that took me a few minutes to do in WebGL took me hours to do in OpenGL. This was the first time I was compiling a library from source and using Visual Studio so there was a bit of learning curve. In the end I ended up using a compiled library to get to my goal but before that I spent a considerable time using Cmake and playing with the properties in Visual Studio to work. 

Who knew a red rectangle could bring such joy :)